Attempt to implement cloning of plugins and setting of ports.
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@3008 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sequencer/rosegardensequencer.h
@@ -195,7 +195,7 @@ public:
     // Set a MappedObject 
     virtual void setMappedProperty(int id,
                                    const QString &property,
-                                   int value);
+                                   float value);
 
     // Get a MappedObject for a type
     //
